move getFeatureAttributes and extractDateValueAttributes to vector_layer.py
